proses proyek

+++ app/(default)/produk/masuk/ui.jsx
@@ -88,8 +88,12 @@ export default function App({ id_produk }) {
   const saveButtonPress = async (onClose) => {
     // if (form.nama == "" || !form.selectKategori.size > 0)
     //   return alert("Nama, dan Kategori wajib diisi!");
+    if (form.jumlah < form.keluar)
+      return alert(
+        `Jumlah tidak boleh kurang dari produk keluar. (Min ${form.keluar})`
+      );
     const res = await fetch(`${apiPath}produkmasuk`, {
-      method,
+      method: form.method,
       headers: {
         "Content-Type": "application/json",
         // 'Content-Type': 'application/x-www-form-urlencoded',
@@ -143,6 +147,8 @@ export default function App({ id_produk }) {
     onOpen();
   };
   const deleteButtonPress = async (data) => {
+    if (data.keluar > 0)
+      return alert("Tidak dapat menghapus. Data keluar tidak kosong.");
     if (confirm("Hapus produk masuk?")) {
       const res = await fetch(`${apiPath}produkmasuk`, {
         method: "DELETE",
@@ -410,11 +416,6 @@ export default function App({ id_produk }) {
     },
   ];
 
-  // if (form.id_kategori) {
-  //   filteredsubkategori = subkategori.data.filter((item) => {
-  //     if (item.id_kategoriproduk == form.id_kategori) return item;
-  //   });
-  // }
   const selectedProduk = produk.data;
   return (
     <div className="flex flex-col">
@@ -435,6 +436,7 @@ export default function App({ id_produk }) {
           Upload Excel
         </Button> */}
       </div>
+      {/* Table Masuk */}
       <Table
         isStriped
         className="pt-3"
@@ -583,7 +585,7 @@ export default function App({ id_produk }) {
                 </div>
                 <Input
                   type="text"
-                  label="jumlah"
+                  label={`Jumlah (Min ${form.keluar}) :`}
                   placeholder="Masukkan jumlah!"
                   value={form.jumlah}
                   onValueChange={(val) => setForm({ ...form, jumlah: val })}
@@ -693,128 +695,6 @@ export default function App({ id_produk }) {
           )}
         </ModalContent>
       </Modal>
-      {/* upload report */}
-      {/* <Modal
-        isOpen={report.isOpen}
-        onOpenChange={report.onOpenChange}
-        scrollBehavior="inside"
-      >
-        <ModalContent>
-          {(onClose) => (
-            <>
-              <ModalHeader className="flex flex-col gap-1">
-                Hasil Upload
-              </ModalHeader>
-              <ModalBody>
-                {reportList.map((r, i) => (
-                  <div key={i}>{r}</div>
-                ))}
-              </ModalBody>
-              <ModalFooter>
-                <Button color="danger" variant="light" onPress={onClose}>
-                  Tutup
-                </Button>
-              </ModalFooter>
-            </>
-          )}
-        </ModalContent>
-      </Modal> */}
-      {/* masuk */}
-      {/* <Modal
-        isOpen={modal.masuk.isOpen}
-        onOpenChange={modal.masuk.onOpenChange}
-        scrollBehavior="inside"
-      >
-        <ModalContent>
-          {(onClose) => (
-            <>
-              <ModalHeader className="flex flex-col gap-1">
-                Produk Masuk
-              </ModalHeader>
-              <ModalBody>
-                <div>
-                  {form.id_kustom} | {form.nama} | {form.nmerek} | {form.tipe} |{" "}
-                  {form.nvendor}{" "}
-                </div>
-                <Input
-                  type="number"
-                  label="Jumlah"
-                  placeholder="Masukkan jumlah!"
-                  value={form.jumlah}
-                  onValueChange={(val) => setForm({ ...form, jumlah: val })}
-                />
-                <Input
-                  type="number"
-                  label="Harga"
-                  placeholder="Masukkan harga!"
-                  value={form.harga}
-                  onValueChange={(val) => setForm({ ...form, harga: val })}
-                />
-                <div className="bg-gray-100 p-3 rounded-lg z-40">
-                  <div>Tanggal</div>
-                  <DatePicker
-                    className="z-40"
-                    placeholderText="Pilih tanggal"
-                    dateFormat="dd/MM/yyyy"
-                    selected={form.startdate}
-                    onChange={(v) =>
-                      setForm({ ...form, startdate: v, tanggal: getDate(v) })
-                    }
-                  />
-                </div>
-                <RadioGroup
-                  orientation="horizontal"
-                  defaultValue={"1"}
-                  value={form.lunas}
-                  onValueChange={(v) => setForm({ ...form, lunas: v })}
-                >
-                  <Radio value="1">Lunas</Radio>
-                  <Radio value="0">Hutang</Radio>
-                </RadioGroup>
-                {form.lunas == "0" ? (
-                  <>
-                    <div className="bg-gray-100 p-3 rounded-lg z-40">
-                      <div>Jatuh Tempo</div>
-                      <DatePicker
-                        className="z-40"
-                        placeholderText="Pilih tanggal"
-                        dateFormat="dd/MM/yyyy"
-                        selected={form.startdateJatuhtempo}
-                        onChange={(v) =>
-                          setForm({
-                            ...form,
-                            startdateJatuhtempo: v,
-                            jatuhtempo: getDate(v),
-                          })
-                        }
-                      />
-                    </div>
-                    <Input
-                      type="number"
-                      label="Terbayar"
-                      placeholder="Masukkan nominal!"
-                      value={form.terbayar}
-                      onValueChange={(val) =>
-                        setForm({ ...form, terbayar: val })
-                      }
-                    />
-                  </>
-                ) : (
-                  <></>
-                )}
-              </ModalBody>
-              <ModalFooter>
-                <Button color="danger" variant="light" onPress={onClose}>
-                  Batal
-                </Button>
-                <Button color="primary" onPress={() => onSimpanClick(onClose)}>
-                  Simpan
-                </Button>
-              </ModalFooter>
-            </>
-          )}
-        </ModalContent>
-      </Modal> */}
     </div>
   );
 }
@@ -823,9 +703,31 @@ const TabelProdukKeluar = ({ id_produk }) => {
   const produkkeluar = useClientFetch(
     `produkkeluar?${id_produk ? `id_produk=${id_produk}` : ""}`
   );
+  const metodepengeluaran = useClientFetch("metodepengeluaran");
   const [page, setPage] = React.useState(1);
+  const [form, setForm] = React.useState({});
   const rowsPerPage = 25;
 
+  const { isOpen, onOpen, onOpenChange } = useDisclosure();
+
+  const saveButtonPress = async (onClose) => {
+    // if (form.nama == "" || !form.selectKategori.size > 0)
+    //   return alert("Nama, dan Kategori wajib diisi!");
+    if (form.selectMetodePengeluaran?.size == 0)
+      return alert(`Metode pengeluaran belum diisi.`);
+    const res = await fetch(`${apiPath}produkkeluar`, {
+      method: form.method,
+      headers: {
+        "Content-Type": "application/json",
+        // 'Content-Type': 'application/x-www-form-urlencoded',
+      },
+      body: JSON.stringify(form),
+    });
+    const json = await res.json();
+    if (res.status == 400) return alert(json.message);
+    onClose();
+    // return alert(json.message);
+  };
   const deleteButtonPress = async (data) => {
     if (confirm("Hapus produk keluar?")) {
       const res = await fetch(`${apiPath}produkkeluar`, {
@@ -841,13 +743,31 @@ const TabelProdukKeluar = ({ id_produk }) => {
       // return alert(await res.json().then((json) => json.message));
     }
   };
+  const editButtonPress = (data) => {
+    setForm({
+      ...data,
+      method: "PUT",
+      modalmode: "Edit",
+      startdate: new Date(data.tanggal),
+      tanggal: getDate(new Date(data.tanggal)),
+      startdateJatuhtempo: data.jatuhtempo ? new Date(data.jatuhtempo) : null,
+      jatuhtempo: data.jatuhtempo ? getDate(new Date(data.jatuhtempo)) : null,
+      oldJumlah: data.jumlah,
+      oldTerbayar: data.terbayar,
+      oldHarga: data.harga,
+      oldSn: data.sn,
+      selectMetodePengeluaran: new Set([data.metodepengeluaran]),
+      lunas: data.jumlah * data.harga > data.terbayar ? "0" : "1",
+    });
+    onOpen();
+  };
 
   const renderCell = React.useCallback((data, columnKey) => {
     const cellValue = data[columnKey];
     switch (columnKey) {
       case "tanggal":
         return getDateFId(new Date(data.tanggal));
-      case "terbayar":
+      case "harga":
         return (
           <div
             className={`text-right px-1 rounded ${
@@ -858,7 +778,7 @@ const TabelProdukKeluar = ({ id_produk }) => {
                 : ""
             }`}
           >
-            <Harga harga={data.terbayar} />
+            <Harga harga={data.harga} />
           </div>
         );
       case "jatuhtempo":
@@ -923,6 +843,8 @@ const TabelProdukKeluar = ({ id_produk }) => {
   const offset = (page - 1) * rowsPerPage;
   if (produkkeluar.error) return <div>failed to load</div>;
   if (produkkeluar.isLoading) return <div>loading...</div>;
+  if (metodepengeluaran.error) return <div>failed to load</div>;
+  if (metodepengeluaran.isLoading) return <div>loading...</div>;
 
   const col = [
     // {
@@ -979,62 +901,171 @@ const TabelProdukKeluar = ({ id_produk }) => {
     },
   ];
 
+  console.log(form.selectMetodePengeluaran?.size);
+
   return (
-    <Table
-      isStriped
-      className="pt-3"
-      aria-label="Example table with custom cells"
-      topContent={
-        <>
-          <div>Masuk</div>
-        </>
-      }
-      bottomContent={
-        pages > 0 ? (
-          <div className="flex w-full justify-center">
-            <Pagination
-              isCompact
-              showControls
-              showShadow
-              color="primary"
-              page={page}
-              total={pages}
-              onChange={(page) => setPage(page)}
-            />
-          </div>
-        ) : null
-      }
-    >
-      <TableHeader columns={col}>
-        {(column) => (
-          <TableColumn
-            key={column.key}
-            align={
-              ["stok", "jumlah", "keluar"].includes(column.key)
-                ? "end"
-                : "start"
-            }
-          >
-            {column.label}
-          </TableColumn>
-        )}
-      </TableHeader>
-      <TableBody
-        items={
-          filteredData ? filteredData.slice(offset, offset + rowsPerPage) : []
+    <>
+      <Table
+        isStriped
+        className="pt-3"
+        aria-label="Example table with custom cells"
+        topContent={
+          <>
+            <div>Keluar</div>
+          </>
         }
-        loadingContent={"Loading..."}
-        emptyContent={"Kosong"}
-        loadingState={loadingState}
+        bottomContent={
+          pages > 0 ? (
+            <div className="flex w-full justify-center">
+              <Pagination
+                isCompact
+                showControls
+                showShadow
+                color="primary"
+                page={page}
+                total={pages}
+                onChange={(page) => setPage(page)}
+              />
+            </div>
+          ) : null
+        }
       >
-        {(item) => (
-          <TableRow key={item.id}>
-            {(columnKey) => (
-              <TableCell>{renderCell(item, columnKey)}</TableCell>
-            )}
-          </TableRow>
-        )}
-      </TableBody>
-    </Table>
+        <TableHeader columns={col}>
+          {(column) => (
+            <TableColumn
+              key={column.key}
+              align={
+                ["stok", "jumlah", "keluar"].includes(column.key)
+                  ? "end"
+                  : "start"
+              }
+            >
+              {column.label}
+            </TableColumn>
+          )}
+        </TableHeader>
+        <TableBody
+          items={
+            filteredData ? filteredData.slice(offset, offset + rowsPerPage) : []
+          }
+          loadingContent={"Loading..."}
+          emptyContent={"Kosong"}
+          loadingState={loadingState}
+        >
+          {(item) => (
+            <TableRow key={item.id}>
+              {(columnKey) => (
+                <TableCell>{renderCell(item, columnKey)}</TableCell>
+              )}
+            </TableRow>
+          )}
+        </TableBody>
+      </Table>
+      <Modal
+        isOpen={isOpen}
+        onOpenChange={onOpenChange}
+        scrollBehavior="inside"
+      >
+        <ModalContent>
+          {(onClose) => (
+            <>
+              <ModalHeader className="flex flex-col gap-1">
+                {form.modalmode} Produk Keluar
+              </ModalHeader>
+              <ModalBody>
+                <div>
+                  {form.produk} | {form.merek} | {form.tipe} | {form.vendor}
+                </div>
+                <Select
+                  required
+                  label="Metode Pengeluaran"
+                  placeholder="Pilih metode"
+                  className="max-w-xs"
+                  selectedKeys={form.selectMetodePengeluaran}
+                  onSelectionChange={(val) =>
+                    setForm({
+                      ...form,
+                      selectMetodePengeluaran: val,
+                      metodepengeluaran: new Set(val).values().next().value,
+                    })
+                  }
+                >
+                  {metodepengeluaran.data.map((item) => (
+                    <SelectItem key={item.nama} value={item.nama}>
+                      {item.nama}
+                    </SelectItem>
+                  ))}
+                </Select>
+                {form.jumlah == 1 ? (
+                  <Input
+                    type="text"
+                    label="S/N"
+                    placeholder="Masukkan S/N!"
+                    value={form.sn}
+                    onValueChange={(val) => setForm({ ...form, sn: val })}
+                  />
+                ) : (
+                  <></>
+                )}
+                <Input
+                  type="number"
+                  label={`Harga (${form.oldHarga})`}
+                  placeholder="Masukkan harga!"
+                  value={form.harga}
+                  onValueChange={(val) => setForm({ ...form, harga: val })}
+                />
+                {/* <Autocomplete
+                  label="Vendor"
+                  variant="bordered"
+                  defaultItems={vendor.data}
+                  placeholder="Cari vendor"
+                  className="max-w-xs"
+                  selectedKey={form.id_vendor}
+                  defaultSelectedKey={form.id_vendor}
+                  defaultInputValue={form.vendor}
+                  onSelectionChange={(v) => setForm({ ...form, id_vendor: v })}
+                >
+                  {(item) => (
+                    <AutocompleteItem key={item.id} textValue={item.nama}>
+                      {item.nama}
+                    </AutocompleteItem>
+                  )}
+                </Autocomplete> */}
+                <div className="bg-gray-100 p-3 rounded-lg z-40">
+                  <div>Tanggal</div>
+                  <DatePicker
+                    className="z-40"
+                    placeholderText="Pilih tanggal"
+                    dateFormat="dd/MM/yyyy"
+                    selected={form.startdate}
+                    onChange={(v) =>
+                      setForm({ ...form, startdate: v, tanggal: getDate(v) })
+                    }
+                  />
+                </div>
+                {/* <Textarea
+                  label="Keterangan"
+                  labelPlacement="inside"
+                  placeholder="Masukkan keterangan! (Opsional)"
+                  value={form.keterangan}
+                  onValueChange={(val) => setForm({ ...form, keterangan: val })}
+                /> */}
+              </ModalBody>
+              <ModalFooter>
+                <Button color="danger" variant="light" onPress={onClose}>
+                  Batal
+                </Button>
+                <Button
+                  color="primary"
+                  onPress={() => saveButtonPress(onClose)}
+                >
+                  Simpan
+                </Button>
+              </ModalFooter>
+            </>
+          )}
+        </ModalContent>
+      </Modal>
+    </>
   );
 };
